requirements.txt
bcrypt==4.0.1
fastapi==0.110.1
openai==1.16.1
passlib==1.7.4
pydantic==2.6.4
pymongo==4.6.3
python-multipart==0.0.9
python-dotenv==1.0.1
python-jose==3.3.0
qdrant-client==1.8.2
uvicorn==0.29.0
